Show table of key-value pairs when clicking generic “kind” tags
